annotation update

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -7,6 +7,6 @@ import typing
 from random import uniform
 
 
-async def wait_random(max_delay: float) -> float:
+async def wait_random(max_delay: int) -> float:
     rand_float = uniform(0, max_delay)
     await asyncio.sleep(rand_float)
